+++ components/Appointment/modal/EditAppointmentModal.tsx
@@ -1,11 +1,10 @@
-import { CalendarIcon, CheckIcon, ChevronDown, Loader2 } from "lucide-react";
+import { CalendarIcon, Loader2 } from "lucide-react";
 import { zodResolver } from "@hookform/resolvers/zod";
 import { useForm } from "react-hook-form";
 import { format } from "date-fns";
 import { ptBR } from "date-fns/locale/pt-BR";
 import { useModal } from "@/hooks/useModalStore";
 import { useEffect, useState } from "react";
-import { Customer } from "@prisma/client";
 import { toast } from "sonner"
 import { useRouter } from "next/navigation";
 
@@ -31,22 +30,12 @@ import {
     FormLabel,
     FormMessage
 } from "@/components/ui/form"
-import {
-    Command,
-    CommandEmpty,
-    CommandGroup,
-    CommandInput,
-    CommandItem
-} from "../../ui/command";
 import { AppointmentSchema, AppointmentSchemaType } from "@/schemas/AppointmentSchema";
 import { formatTime } from "@/utils/globalFunctions";
 import { updateAppointment } from "@/services/AppointmentService";
-import { listCustomers } from "@/services/CustomerService";
+import PopoverCustomers from "@/components/PopoverCustomers";
 
 const EditAppointmentModal = () => {
-    const [customers, setCustomers] = useState<Customer[]>([]);
-    const [customersLoading, setCustomersLoading] = useState<boolean>(false);
-    const [openCustomerPopover, setOpenCustomerPopover] = useState<boolean>(false);
     const [openDatePopover, setOpenDatePopover] = useState<boolean>(false);
 
     const router = useRouter();
@@ -68,8 +57,6 @@ const EditAppointmentModal = () => {
 
     useEffect(() => {
         if (appointmentData) {
-            loadCustomers(true);
-
             form.setValue('customer', appointmentData.customerId);
             form.setValue('date', appointmentData.date);
             form.setValue('description', appointmentData.description);
@@ -98,33 +85,10 @@ const EditAppointmentModal = () => {
         onClose();
     }
 
-    async function loadCustomers(isOpen: boolean) {
-        try {
-            if (isOpen) {
-                setCustomers(new Array());
-                setCustomersLoading(true);
-
-                const customersArray = await listCustomers();
-
-                setCustomers(customersArray);
-            }
-        } catch (error) {
-            console.log('Falha ao consultar clientes - ', error);
-            toast.error('Falha ao consultar clientes. Por favor tente novamente.');
-        }
-
-        setCustomersLoading(false);
-    }
-
     function getDateWhithoutHours() {
         const dateNow = new Date();
         const dateString = `${dateNow.getFullYear()}-${dateNow.getMonth() + 1}-${dateNow.getDate()}`;
         return new Date(dateString);
-    }
-
-    function handleOpenCustomerPopover(open: boolean) {
-        loadCustomers(open);
-        setOpenCustomerPopover(open);
     }
 
     return (
@@ -147,63 +111,7 @@ const EditAppointmentModal = () => {
                                 <FormItem className="flex flex-col">
                                     <FormLabel>Cliente</FormLabel>
 
-                                    <Popover onOpenChange={handleOpenCustomerPopover} open={openCustomerPopover}>
-                                        <PopoverTrigger asChild>
-                                            <FormControl>
-                                                <Button
-                                                    variant='outline'
-                                                    role="combobox"
-                                                    className={cn(
-                                                        "w-full justify-between",
-                                                        !field.value && "text-muted-foreground"
-                                                    )}
-                                                    onClick={() => setOpenCustomerPopover(openCustomerPopover => !openCustomerPopover)}
-                                                >
-                                                    {field.value
-                                                        ? customers?.find((customer) => customer.id === field.value)?.name
-                                                        : 'Selecione um cliente'}
-
-
-                                                    <ChevronDown className="ml-2 h-4 w-4 shrink-0 opacity-50" />
-                                                </Button>
-                                            </FormControl>
-                                        </PopoverTrigger>
-
-                                        <PopoverContent className="p-0 PopoverContent">
-                                            <Command>
-                                                <CommandInput
-                                                    placeholder="Digite para pesquisar..."
-                                                    className="h-9"
-                                                />
-
-                                                <CommandEmpty>{customersLoading ? 'Carregando...' : 'Nenhum cliente encontrado.'}</CommandEmpty>
-
-                                                <CommandGroup>
-                                                    {customers?.map((customer) => (
-                                                        <CommandItem
-                                                            value={customer.name}
-                                                            key={customer.id}
-                                                            onSelect={() => {
-                                                                form.setValue('customer', customer.id);
-                                                                form.trigger('customer');
-                                                                setOpenCustomerPopover(openCustomerPopover => !openCustomerPopover);
-                                                            }}
-                                                        >
-                                                            {customer.name}
-                                                            <CheckIcon
-                                                                className={cn(
-                                                                    "ml-auto h-4 w-4",
-                                                                    customer.id === field.value
-                                                                        ? "opacity-100"
-                                                                        : "opacity-0"
-                                                                )}
-                                                            />
-                                                        </CommandItem>
-                                                    ))}
-                                                </CommandGroup>
-                                            </Command>
-                                        </PopoverContent>
-                                    </Popover>
+                                    <PopoverCustomers field={field} form={form} />
 
                                     <FormMessage />
                                 </FormItem>
